refactor(shared): move verbose timing helpers into reporting
Relocated src/shared/src/utils/time.js to src/shared/src/reporting/time.js because the helpers orchestrate verbose logging/reporting rather than generic utilities. Updated re-exports, CLI imports, and tests to follow the new home.
