+++ src/shared/src/reporting/index.js
@@ -1,1 +1,6 @@
 export { createMetricsTracker } from "./metrics.js";
+export {
+    createVerboseDurationLogger,
+    formatDuration,
+    timeSync
+} from "./time.js";
