+++ src/shared/src/reporting/time.js
