+++ src/shared/src/utils/index.js
@@ -13,4 +13,3 @@ export * from "./numeric-options.js";
 export * from "./object.js";
 export * from "./regexp.js";
 export * from "./string.js";
-export * from "./time.js";
